Compare every single item by mtime
The 500x speedup in https://github.com/gabebw/rust-lister/pull/1 was
because the lister was doing a lot less work: by doing `take(n)`, we
never needed to build up a giant (700,000+ item) Vec, then sort it.

But in order to compare all the items, we need to get all of their file
metadata, which takes time.

Now the program works correctly, but is back to being slow.